brandfetch/requirements.txt
requests
supabase
python-dotenv
websockets
google-api-python-client
google-auth
